refactor(activerecord): move Mysql2Adapter + migration errors to their Rails paths
Followup to the api-compare extractor work: Rails fidelity wants
these classes to LIVE where Rails puts them, not be re-exported
from there. Pre-launch, so no backward-compat shims needed.

- Move Mysql2Adapter: adapters/mysql2-adapter.ts →
  connection-adapters/mysql2-adapter.ts. Rails has it at
  activerecord/lib/active_record/connection_adapters/mysql2_adapter.rb.
  The test moves with it. adapters/index.ts re-exports from the
  new location so the adapters/ barrel still works. Package.json
  subpath export './adapters/mysql2-adapter.js' removed —
  './connection-adapters/mysql2-adapter.js' is canonical. Internal
  dynamic imports in tasks/database-tasks, tasks/mysql-database-tasks,
  test-adapter, connection-adapters.ts, establish-connection.test,
  and trailties/database.ts updated.
- Move the 13 migration error classes: migration-errors.ts back
  into migration.ts where Rails defines them. migration-errors.ts
  deleted. The split was originally for circular-dep avoidance;
  the ESM cycle (migration.ts → internal-metadata.ts → migration.ts
  for EnvironmentStorageError) is safe because the error is only
  referenced from a method body, not at module init time.
- Copilot review #2 on #654: POSIX-normalize relPath at creation so
  manifest keys are platform-stable. resolveRelModule now uses
  path.posix.* consistently and is exported for direct testing.
- Copilot review #1 on #654: 6 focused unit tests for
  resolveRelModule. Extractor's main() gated on `if __main__`
  so importing it in the test doesn't trigger a fresh extraction.
- Cache gate in extract-ruby-api.rb: skip Ripper pass (~4-5s) when
  rails-api.json is newer than the Rails source HEAD. Honours
  API_COMPARE_FORCE=1. Rails is pinned to a tag by fetch-rails.sh,
  so output never changes between runs.
