+++ packages/activerecord/src/establish-connection.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import { Base } from "./base.js";
 import { PostgreSQLAdapter } from "./connection-adapters/postgresql-adapter.js";
 import { SQLite3Adapter } from "./connection-adapters/sqlite3-adapter.js";
-import { Mysql2Adapter } from "./adapters/mysql2-adapter.js";
+import { Mysql2Adapter } from "./connection-adapters/mysql2-adapter.js";
 import { ConnectionHandler } from "./connection-adapters/abstract/connection-handler.js";
 import { writeFileSync, mkdirSync, rmSync } from "fs";
 import { join } from "path";
