+++ packages/activerecord/src/internal-metadata.ts
@@ -7,7 +7,7 @@
 import type { DatabaseAdapter } from "./adapter.js";
 import { detectAdapterName } from "./adapter-name.js";
 import { quoteIdentifier, quoteTableName } from "./connection-adapters/abstract/quoting.js";
-import { EnvironmentStorageError } from "./migration-errors.js";
+import { EnvironmentStorageError } from "./migration.js";
 import {
   Table,
   SelectManager,
@@ -133,7 +133,7 @@ export class InternalMetadata {
       // internal_metadata is disabled; surface the same error here so
       // callers that attempt to write through a disabled instance fail
       // loudly rather than silently no-op. Imported statically from
-      // migration-errors.ts to avoid a circular dep with migration.ts.
+      // migration.ts (the cycle is ESM-safe because EnvironmentStorageError is only used in method bodies).
       throw new EnvironmentStorageError();
     }
     const existing = await this.selectEntry(key);
