+++ packages/activerecord/src/tasks/database-tasks.ts
@@ -1052,7 +1052,7 @@ export class DatabaseTasks {
       });
     }
     if (/mysql|trilogy/.test(adapter)) {
-      const { Mysql2Adapter } = await import("../adapters/mysql2-adapter.js");
+      const { Mysql2Adapter } = await import("../connection-adapters/mysql2-adapter.js");
       const c = config.configuration;
       if (c.url) return new Mysql2Adapter(String(c.url));
       return new Mysql2Adapter({
